server/requirements.txt
flask === 2.2.2
flask_cors === 3.0.10
torch
openai-whisper
matplotlib ===  3.4.1
scipy === 1.10.1
tokenizers === 0.13.3
transformers === 4.24.0